Move ChainId to ledger::ethereum module
ChainId is the property of the Ethereum
 ledger for swap protocols. No need to have
 it in the ethereum_support crate.
