+++ cnd/src/comit_api/mod.rs
@@ -146,7 +146,7 @@ impl FromHeader for Decision {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::swap_protocols::HashFunction;
+    use crate::swap_protocols::{ledger::ethereum, HashFunction};
     use bitcoin::Amount;
     use ethereum_support::{Address, Erc20Quantity, U256};
     use spectral::prelude::*;
@@ -224,7 +224,7 @@ mod tests {
 
     #[test]
     fn ethereum_ledger_to_header() {
-        let ledger = LedgerKind::Ethereum(Ethereum::new(ethereum_support::ChainId::ropsten()));
+        let ledger = LedgerKind::Ethereum(Ethereum::new(ethereum::ChainId::ropsten()));
         let header = ledger.to_header().unwrap();
 
         assert_eq!(
